+++ module/sdm.mjs
@@ -55,7 +55,6 @@ import { $l10n } from './helpers/globalUtils.mjs';
 const { ActiveEffectConfig } = foundry.applications.sheets;
 const { Actors, Items } = foundry.documents.collections;
 const { DocumentSheetConfig } = foundry.applications.apps;
-const sheets = foundry.appv1.sheets;
 
 /* -------------------------------------------- */
 /*  Init Hook                                   */
@@ -146,10 +145,15 @@ Hooks.on('getSceneControlButtons', function (controls) {
       if (active) await game.sdm.api.player.diceOracles();
     }
   };
-});
-
-Hooks.on('renderSidebarTab', (app, html) => {
-  if (!game.user.isGM) return;
+  controls.tokens.tools['sdm-burden-generator'] = {
+    icon: 'fa-solid fa-face-dizzy',
+    name: 'sdm-burden-generator',
+    title: 'SDM.BurdenGenerator.Title',
+    button: true,
+    onChange: async (event, active) => {
+      if (active) await game.sdm.api.gm.openBurdenGeneratorDialog();
+    }
+  };
 });
 
 Hooks.on('renderActorDirectory', (app, html) => {
@@ -167,6 +171,19 @@ Hooks.on('renderActorDirectory', (app, html) => {
 
   html.querySelector('.random-pc').addEventListener('click', ev => {
     game.sdm.api.gm.characterGeneratorDialog();
+  });
+});
+
+Hooks.on('renderItemDirectory', (app, html) => {
+  if (!game.user.isGM) return;
+
+  html.querySelector('.directory-header .create-entry').insertAdjacentHTML(
+    'beforebegin',
+    `<button type="button" class="burden-generator" style="flex-basis: 100%"><i class="fa-solid fa-face-dizzy" inert></i><span>${$l10n('SDM.BurdenGenerator.Title')}</span></button>`
+  );
+
+  html.querySelector('.burden-generator').addEventListener('click', ev => {
+    game.sdm.api.gm.openBurdenGeneratorDialog();
   });
 });
 
@@ -190,9 +207,11 @@ Hooks.on('renderCombatTracker', (app, html) => {
 Hooks.on('updateCombat', async (combat, update) => {
   if (!game.user.isGM) return;
 
+  const reroll = game.settings.get('sdm', 'rerollInitiativeEveryRound');
+  if (!reroll) return;
+
   if (update && update.round && update.round > 1) {
-    const reroll = game.settings.get('sdm', 'rerollInitiativeEveryRound');
-    if (reroll) await game.sdm.api.gm.groupInitiative({ reroll: true });
+    await game.sdm.api.gm.groupInitiative({ reroll: true });
   }
 });
 
@@ -502,7 +521,6 @@ Hooks.once('init', function () {
     trait: models.SdmTrait,
     burden: models.SdmBurden
   };
-
 
   _configureFonts();